backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Marshmallow
marshmallow-sqlalchemy
Flask-JWT-Extended
Flask-Cors
python-dotenv
werkzeug
